Add checking directory structure of the project (#6)
* Add checking directory structure of the project

* Ignore tests/__pycache__

* Add test data: simple project

This test data has only an empty manifest file and an empty ledger file.
This is a simple structure of projects.

simple-project/
├── Largo.toml
└── book/
     └── 2021.ledger

* Add test code for `Project` instantiation

Two test cases are added.

1. `simple-project` which has correct structure.
2. test with an empty `tmpdir`, which fails.

* Add project structure checking function

The class `Project` checks the structure of the given project when it is
instantiated. It checks the following items:
- Whether the given manifest file exists,
- Whether the `book` directory exists in the project root.

If the check failed, the constructor of `Project` throws an exception.
The execption `StructureError` is that.