Add reading account names from the given manifest file (#7)
* Add reading account names from the given manifest file

* Add toml package

* Add loading a manifest file in instantiation

* Add test code and test data

Added test data are the following:
- `tests/data/simple-project` : English account names
- `tests/data/japanese-manifest` : Japanese account names

The function `test_account_names` in `tests/test_project.py` is the test
code for the data above.

* Ignore *.pyc files

Formerly, .gitignore ignored __pycache__ directories.
But now it ignores all pycache files.

* Add properties for account names to `Project`

The class `Project` has a property `account` which returns an instance
of the class `AccountNames`. `AccountNames` is a wrapper for a
dictionary which provides properties.

Now the test fails because `tests/data/japanese-manifest` lacks `book`
directory.

* Add book directory in `tests/data/japanese-manifest`

* Add error test case for referencing account names

The test data is an empty manifest file in `empty-manifest`