Add `largo bs` command which shows the balance sheet of a year (#2)
* `largo bs` command shows the balance sheet of a year

Issue #1

* Add CLI boilerplate (#4)

* Add CLI boiler plate

* Add CLI boilerplate

- `__cli__.py` implements the entrypoint which initializes `Application`
  instance.
- `bs_command.py` implements a boilerplate of `bs` command.
- `pyproject.toml` defines the `largo` command.
- `.gitignore` ignores `__pycache__` directory.

* Add an option to specify the manifest file for development (#5)

* Add an option to specify the manifest file for development

* Add `--manifest-path` option to `largo bs` command.

It takes a value. The default value is `Largo.toml`.
The `bs` command prints the value of the option.

* Add class Project which represents a project and contains the manifest path.

Now it has only a instance variable `manifest_path`.
The `bs` command prints the project variable out.

* Add checking directory structure of the project (#6)

* Add checking directory structure of the project

* Ignore tests/__pycache__

* Add test data: simple project

This test data has only an empty manifest file and an empty ledger file.
This is a simple structure of projects.

simple-project/
├── Largo.toml
└── book/
     └── 2021.ledger

* Add test code for `Project` instantiation

Two test cases are added.

1. `simple-project` which has correct structure.
2. test with an empty `tmpdir`, which fails.

* Add project structure checking function

The class `Project` checks the structure of the given project when it is
instantiated. It checks the following items:
- Whether the given manifest file exists,
- Whether the `book` directory exists in the project root.

If the check failed, the constructor of `Project` throws an exception.
The execption `StructureError` is that.

* Add reading account names from the given manifest file (#7)

* Add reading account names from the given manifest file

* Add toml package

* Add loading a manifest file in instantiation

* Add test code and test data

Added test data are the following:
- `tests/data/simple-project` : English account names
- `tests/data/japanese-manifest` : Japanese account names

The function `test_account_names` in `tests/test_project.py` is the test
code for the data above.

* Ignore *.pyc files

Formerly, .gitignore ignored __pycache__ directories.
But now it ignores all pycache files.

* Add properties for account names to `Project`

The class `Project` has a property `account` which returns an instance
of the class `AccountNames`. `AccountNames` is a wrapper for a
dictionary which provides properties.

Now the test fails because `tests/data/japanese-manifest` lacks `book`
directory.

* Add book directory in `tests/data/japanese-manifest`

* Add error test case for referencing account names

The test data is an empty manifest file in `empty-manifest`

* Add reading the current year ledger file (#8)

* Add reading the current year ledger file

* Add balance_sheet module and BalanceSheet class

It only has the constructor, which takes an instance of Project.
The test code initializes the BalanceSheet class.

* Add `book` method to class `Project`

This method looks up the book file of the specified year.

* Add `command_arguments` property to `BalanceSheet`

Test codes are also added.

* Add ledger-cli command invocation (#9)

* Add ledger-cli command invocation

It invokes the ledger-cli and pass the input through stdin.

* Add showing the balance sheet of the current year

* Add handling of the year argument

The command `largo bs` accepts the year of the balance sheet.

* Rename and refactor `AccountNames` (#10)

* Rename and refactor `AcountNames`.

The name of the class `AccountNames` is renamed to `Account`. It is
refactored with `dataclass` and check if the fields exists in the
given manifest.

* Rename `AccountName` to `Account`

* Rewrite class `Account` with dataclass

* Refactor fixtures in test codes (#11)

* Refactor fixtures in test codes

Fixtures with datadir are refactored. The factored ones are written in
_conftest.py_.

* Refactor largo project fixtures in _conftest.py_

* Ignore _dist/_ directory