Change the current year BS to the latest year one (#23)
* Change the current year BS to the latest year one

Now `largo bs` shows the current year BS.
For convenience, I want to change `largo bs` to show the latest year
one.

* Add `latest_year` method to `class Project`

The method `latest_year` returns the latest year of the ledger books in
the project directory. If there are book files with non-numeric stems,
it raises an exception because the stem cannot be converted into `int`.

The following test data is used:
  - `simple-project`
  - `non-numeric-books`

* Change to use `latest_year` method in `bs_command.py`

* Correct type hint for `latest_year`