Refactor `DateRange` handling codes in _largo/pl_command.py_
In order to implement period arguments handling for `largo bs` and
`largo cf` (in issues #29, #30), `DateRange` handling code in
`PlCommand` is refactored out.

- [ ] Refactor code which converts arguments into year and month pair.
- [ ] Modify `LedgerInvoker.build` to take a `DateRange` object.

The converter of arguments to pairs takes the default year. It returns
a object of `DateRange`.
The class `DateRange` must have a method which returns its year.