Refactor `DateRange` handling codes in _largo/pl_command.py_ (#31)
* Refactor `DateRange` handling codes in _largo/pl_command.py_

In order to implement period arguments handling for `largo bs` and
`largo cf` (in issues #29, #30), `DateRange` handling code in
`PlCommand` is refactored out.

- [ ] Refactor code which converts arguments into year and month pair.
- [ ] Modify `LedgerInvoker.build` to take a `DateRange` object.

The converter of arguments to pairs takes the default year. It returns
a object of `DateRange`.
The class `DateRange` must have a method which returns its year.

* Suppress pyright error of type mismatch by introducing typing.cast

* Extract the function of `date_argument` to `month, year`

* Update python and pytest

Python 3.10 is installed and pytest is udpated to 6.2

* Rewrite with assignment operator

* Add classes which represent year, month, and year/month pair.

* Add year/month argument parser which returns `Year | Month | YearMonth`

* Add doctest for `DateRange`

* Support month in int type in `__init__` of `DateRange`

* Add function `date_argument_to_date_range`

* Refactor out date argument to date range routine

* Remove _test_pl_command.py_

Because it is now empty.

* Add `year` property to `LedgerInvoker`

The class `ProfitLoss` and `BlanceSheet` have their year. The `build`
method uses the `year` property to look a ledger file up.